refactor(cli): hoist getDb() out of preferences
extracts the cached YapprDb opener + legacy settings.json import into
lib/db.ts so chat persistence (next commit) can reuse the same handle.
no behavior change.
